requirements.txt
python-dotenv
requests
bs4
pydantic
lxml